core: rename UIComponents
